Python: Fixup CWE-094 tests
