--- python/ql/test/experimental/query-tests/Security-new-dataflow/CWE-094/CodeInjection.qlref
@@ -1,1 +1,0 @@
-experimental/Security-new-dataflow/CWE-094/CodeInjection.ql
