--- python/ql/test/experimental/query-tests/Security-new-dataflow/CWE-094/code_injection.py
@@ -1,10 +1,0 @@
-from flask import Flask, request
-app = Flask(__name__)
-
-@app.route("/code-execution")
-def code_execution():
-    code = request.args.get("code")
-    exec(code)
-    eval(code)
-    cmd = compile(code, "<filename>", "exec")
-    exec(cmd)
