--- python/ql/test/query-tests/Security/CWE-094/options
@@ -1,1 +1,0 @@
-semmle-extractor-options: -p ../lib/ --max-import-depth=3
